+++ src/media/lua/client/AoqiaCarwannaExtendedClient/ui/tooltip.lua
@@ -1,14 +1,6 @@
--- -------------------------------------------------------------------------- --
---                          Vehicle UI tooltip stuff                          --
--- -------------------------------------------------------------------------- --
-
--- AoqiaCarwannaExtended requires.
+local aoqia_table = require("AoqiaZomboidUtilsShared/table")
 local mod_constants = require("AoqiaCarwannaExtendedShared/mod_constants")
 
--- std globals
-local tostring = tostring
-
--- TIS globals.
 local getCore = getCore
 local getMouseX = getMouseX
 local getMouseY = getMouseY
@@ -23,6 +15,7 @@ local logger = mod_constants.LOGGER
 
 local tooltip = {}
 
+--- @type function
 tooltip.o_tooltip_render = nil
 
 -- Font stuff
@@ -67,8 +60,8 @@ function tooltip.modified_render(self, hard_width)
     self.item:DoTooltip(self.tooltip)
     self.tooltip:setMeasureOnly(false)
 
-    local tw = self.tooltip:getWidth()
-    local th = self.tooltip:getHeight()
+    local tw = self.tooltip:getWidth() --[[@as number]]
+    local th = self.tooltip:getHeight() --[[@as number]]
 
     local v1 = mx + 11
     local v2 = (screen_width - tw) - 1
@@ -123,63 +116,60 @@ end
 --- My hooked render function
 --- @param self ISToolTipInv
 function tooltip.render(self)
-    if tooltip.o_tooltip_render == nil or (ISContextMenu.instance
-        and ISContextMenu.instance.visibleCheck) then
+    if tooltip.o_tooltip_render == nil
+    or (ISContextMenu.instance and ISContextMenu.instance.visibleCheck) then
         return
     end
 
-    --- @type InventoryItem
-    local item = self.item
+    local item = self.item --[[@as InventoryItem | nil]]
     if item == nil or item:getFullType() ~= (mod_constants.MOD_ID .. ".Pinkslip") then
         tooltip.o_tooltip_render(self)
         return
     end
 
-    local sbvars = SandboxVars[mod_constants.MOD_ID] --[[@as SandboxVarsDummy]]
-    local mdata = item:getModData() --[[@as ModDataDummy]]
-
-    local text = getText(("IGUI_%s_Tooltip"):format(mod_constants.MOD_ID)) --[[@as string]]
-
-    text = text
-        .. " <LINE> "
-        .. getText(("IGUI_%s_VehicleSkin"):format(mod_constants.MOD_ID), mdata.Skin)
-        .. " <LINE> "
-        .. getText(("IGUI_%s_HasKey"):format(mod_constants.MOD_ID),
-            mdata.HasKey and "Yes" or "No")
-        .. " <LINE> "
-        .. getText(("IGUI_%s_Hotwired"):format(mod_constants.MOD_ID),
-            mdata.Hotwired and "Yes" or "No")
-
-    local parts = mdata.Parts
-    if sbvars.DoShowAllParts and parts then
-        for i = 1, #parts.index do
-            local part = parts.index[i]
-            local data = parts.values[i]
-
-            text = text
-                .. ("%s (%s)%% <LINE> "):format(
-                    getText("IGUI_VehiclePart" .. part),
-                    tostring(data.Condition))
-        end
-    else
-        text = text
-            .. getText(("IGUI_%s_PartsMissing"):format(mod_constants.MOD_ID), mdata.PartsMissing)
-            .. getText(("IGUI_%s_PartsDamaged"):format(mod_constants.MOD_ID), mdata.PartsDamaged)
+    local mdata = aoqia_table.init_mdata(item, mod_constants.MOD_ID) --[[@as ModDataDummy | nil]]
+    if mdata == nil or mdata.Name == nil then
+        tooltip.o_tooltip_render(self)
+        return
     end
 
-    local v1 = text_manager:MeasureStringX(font_type, item:getName())
+    local text = getText(("IGUI_%s_VehicleName"):format(mod_constants.MOD_ID), mdata.Name or "nil")
+        .. "\n"
+        .. getText(("IGUI_%s_VehicleSkin"):format(mod_constants.MOD_ID), mdata.Skin or "nil")
+        .. "\n"
+        .. getText(("IGUI_%s_HasKey"):format(mod_constants.MOD_ID), mdata.HasKey and "Yes" or "No")
+        .. "\n"
+        .. getText(("IGUI_%s_Hotwired"):format(mod_constants.MOD_ID),
+            mdata.Hotwired and "Yes" or "No")
+        .. "\n"
+        .. getText(("IGUI_%s_PartsMissing"):format(mod_constants.MOD_ID), mdata.PartsMissing)
+        .. "\n"
+        .. getText(("IGUI_%s_PartsDamaged"):format(mod_constants.MOD_ID), mdata.PartsDamaged)
+
+    -- local parts = mdata.Parts
+    -- if sbvars.DoShowAllParts and parts then
+    --     for i = 1, #parts.index do
+    --         local part = parts.index[i]
+    --         local data = parts.values[i]
+
+    --         text = text
+    --             .. ("%s (%s)%% <LINE> "):format(
+    --                 getText("IGUI_VehiclePart" .. part),
+    --                 tostring(data.Condition))
+    --     end
+    -- else
+    --     text = text
+    --         .. getText(("IGUI_%s_PartsMissing"):format(mod_constants.MOD_ID), mdata.PartsMissing)
+    --         .. getText(("IGUI_%s_PartsDamaged"):format(mod_constants.MOD_ID), mdata.PartsDamaged)
+    -- end
+
+    local v1 = self.tooltip:getWidth()
     local v2 = text_manager:MeasureStringX(font_type, text)
     local text_width = v1 > v2 and v1 or v2
 
-    local text_height = text_manager:MeasureStringY(font_type, mdata.Name)
-
-    if text then
-        text_height = text_height + text_manager:MeasureStringY(font_type, text) + 8
-    end
-
+    local text_height = text_manager:MeasureStringY(font_type, text) + 8
     local tooltip_width = text_width + font_bounds
     tooltip.modified_render(self, tooltip_width)
-
     local tooltip_height = self.tooltip:getHeight() - 1
     self:setX(self.tooltip:getX() - 11)
 
@@ -193,14 +183,11 @@ function tooltip.render(self)
     local col_background = self.backgroundColor
     local col_border = self.borderColor
 
-    v1 = col_background.a + 0.4
-    local alpha = v1 < 1 and v1 or 1
-
     self:drawRect(0,
         tooltip_height,
         tooltip_width,
         text_height + 8,
-        alpha,
+        col_background.a,
         col_background.r,
         col_background.g,
         col_background.b)
@@ -215,25 +202,17 @@ function tooltip.render(self)
         col_border.b)
 
     local line_height = text_manager:getFontFromEnum(font_type):getLineHeight()
-    local col_font = { a = 1.0, r = 0.9, g = 0.9, b = 0.9 }
+    local col_font = { r = 0.9, g = 0.9, b = 0.9, a = 1.0 }
 
-    local y = (y_offset + (15 - line_height) / 2)
-    self:drawText(mdata.Name, x_offset, y, col_font.r, col_font.g, col_font.b,
-        col_font.a, font_type)
-
-    if text then
-        y = y + (line_height * 1.5)
-        self:drawText(text,
-            x + 1,
-            y,
-            col_font.r,
-            col_font.g,
-            col_font.b,
-            col_font.a,
-            font_type)
-    end
-
-    y_offset = y_offset + 12
+    y_offset = (y_offset + (15 - line_height) / 2)
+    self:drawText(text,
+        x_offset,
+        y_offset,
+        col_font.r,
+        col_font.g,
+        col_font.b,
+        col_font.a,
+        font_type)
 end
 
 return tooltip