+++ src/media/lua/shared/AoqiaCarwannaExtendedShared/mod_constants.lua
@@ -5,7 +5,7 @@ local logger = require("AoqiaZomboidUtilsShared/logger")
 local mod_constants = {}
 
 mod_constants.MOD_ID = "AoqiaCarwannaExtended"
-mod_constants.MOD_VERSION = "0.4.2"
+mod_constants.MOD_VERSION = "1.0.1"
 
 mod_constants.LOGGER = logger:new(mod_constants.MOD_ID)
 
@@ -93,6 +93,9 @@ mod_constants.FUELTANK_NAMES = { "500FuelTank", "1000FuelTank" }
 --- Safehouse
 --- @field DoSafehouseOnly boolean
 --- @field SafehouseDistance int
+--- Parking Meter
+--- @field DoParkingMeterOnly boolean
+--- @field ParkingMeterDistance int
 --- Zombie Loot
 --- @field DoZedLoot boolean
 --- @field ZedLootChance float