+++ CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-02-18
+
+### Fixed
+
+- Not checking if player is near a parking meter for using pinkslips.
+
+## [1.0.0] - 2025-02-12
+
+### Added
+
+-   Pinkslip item hover details
+-   Parking meter tile for creating and using pinkslips
+
 ## [0.4.2] - 2024-12-01
 
 ### Fixed
 
-- Vehicle spawning above level 0 (engine limitation).
+-   Vehicle spawning above level 0 (engine limitation).
 
 ## [0.4.1] - 2024-11-21